move store initialization to conversations/init.ts
