+++ src/conversations/init.ts
@@ -2,7 +2,7 @@
  * Conversation Store Initialization
  *
  * Sets up the Redux store with saga middleware, IndexedDB persistence,
- * and returns a ConversationStore.
+ * and provides singleton management for ConversationStore.
  *
  * This module handles:
  * 1. IndexedDB polyfill initialization (must happen first)
@@ -10,12 +10,16 @@
  * 3. Redux store setup with saga middleware
  * 4. Root saga startup
  * 5. Waiting for IDB data to load into Redux
+ * 6. KeyManager initialization
+ * 7. Singleton management
  */
 
 import createSagaMiddleware from 'redux-saga';
 
 import { logger } from '../app/logger.js';
-import type { SpaceId } from './types.js';
+import type { SpaceId, InitializeStoreOptions } from './types.js';
+import { getKeyManager } from './key-manager.js';
+import { ConversationStore } from './store.js';
 
 import { DbApi } from '@lumo/indexedDb/db.js';
 import { generateSpaceKeyBase64 } from '@lumo/crypto/index.js';
@@ -35,9 +39,124 @@ import { setupStore, type LumoSagaContext, type LumoStore } from '@lumo/redux/st
 import { LumoApi } from '@lumo/remote/api.js';
 import type { Space } from '@lumo/types.js';
 
-import { ConversationStore } from './store.js';
+// ============================================================================
+// Singleton Management
+// ============================================================================
 
-export interface StoreConfig {
+let activeStore: ConversationStore | null = null;
+
+/**
+ * Get the active conversation store
+ *
+ * Returns the initialized store, or undefined if no store is available.
+ * Callers should handle undefined gracefully (stateless mode).
+ */
+export function getConversationStore(): ConversationStore | undefined {
+    return activeStore ?? undefined;
+}
+
+/**
+ * Set the active conversation store (for mock mode or CLI fallback)
+ */
+export function setConversationStore(store: ConversationStore): void {
+    activeStore = store;
+}
+
+/**
+ * Reset the conversation store (for testing)
+ */
+export function resetConversationStore(): void {
+    activeStore = null;
+}
+
+// ============================================================================
+// High-Level Initialization
+// ============================================================================
+
+/**
+ * Initialize the conversation store
+ *
+ * Creates the ConversationStore (Redux + IndexedDB) if possible.
+ * Logs warnings if initialization fails - callers should handle this
+ * gracefully (server works stateless, CLI uses local Turn array).
+ *
+ * Requires:
+ * - Auth provider supports persistence (has cached encryption keys)
+ * - keyPassword is available (for master key decryption)
+ */
+export async function initializeConversationStore(
+    options: InitializeStoreOptions
+): Promise<void> {
+    const { authProvider, conversationsConfig } = options;
+
+    // Check if store is disabled via config
+    if (!conversationsConfig.enableStore) {
+        logger.info('ConversationStore disabled via config');
+        return;
+    }
+
+    // Check if ConversationStore can be used
+    const storeWarning = authProvider.getConversationStoreWarning();
+    if (storeWarning) {
+        logger.warn({ method: authProvider.method }, storeWarning);
+        return;
+    }
+
+    // If we get here, getConversationStoreWarning() confirmed keyPassword exists
+    const keyPassword = authProvider.getKeyPassword()!;
+
+    // Get cached keys from browser provider if available
+    const cachedUserKeys = authProvider.getCachedUserKeys?.();
+    const cachedMasterKeys = authProvider.getCachedMasterKeys?.();
+
+    logger.info(
+        {
+            method: authProvider.method,
+            hasCachedUserKeys: !!cachedUserKeys,
+            hasCachedMasterKeys: !!cachedMasterKeys,
+        },
+        'Initializing KeyManager...'
+    );
+
+    // Initialize KeyManager
+    const keyManager = getKeyManager({
+        protonApi: options.protonApi,
+        cachedUserKeys,
+        cachedMasterKeys,
+    });
+
+    try {
+        await keyManager.initialize(keyPassword);
+
+        // Get master key as base64 for crypto layer
+        const masterKeyBase64 = keyManager.getMasterKeyBase64();
+
+        const result = await createReduxStore({
+            sessionUid: options.uid,
+            userId: authProvider.getUserId() ?? options.uid,
+            masterKey: masterKeyBase64,
+            projectName: conversationsConfig.projectName,
+        });
+
+        activeStore = result.conversationStore;
+        logger.info('ConversationStore initialized');
+
+        // Pull incomplete conversations in background when sync is enabled
+        if (conversationsConfig.enableSync) {
+            pullIncompleteConversations(result.store, result.spaceId)
+                .catch(err => logger.error({ error: err }, 'Failed to pull incomplete conversations'));
+        }
+    } catch (error) {
+        const msg = error instanceof Error ? error.message : String(error);
+        logger.error({ error: msg }, 'Failed to initialize store. Continuing without store.');
+    }
+}
+
+// ============================================================================
+// Redux Store Setup (Internal)
+// ============================================================================
+
+interface ReduxStoreConfig {
     /** Session UID for API authentication (x-pm-uid header) */
     sessionUid: string;
     /** Stable user ID for database naming (userKeys[0].ID) */
@@ -47,7 +166,7 @@ export interface StoreConfig {
     projectName: string;
 }
 
-export interface StoreResult {
+interface ReduxStoreResult {
     store: LumoStore;
     conversationStore: ConversationStore;
     dbApi: DbApi;
@@ -55,7 +174,7 @@ export interface StoreResult {
 }
 
 /**
- * Initialize the upstream storage system
+ * Create the Redux-backed store infrastructure
  *
  * This sets up:
  * - IndexedDB (via indexeddbshim) for local persistence
@@ -67,9 +186,9 @@ export interface StoreResult {
  * 1. Find existing space by projectName in Redux state
  * 2. Create new space with projectName if no match
  */
-export async function initializeStore(
-    config: StoreConfig
-): Promise<StoreResult> {
+async function createReduxStore(
+    config: ReduxStoreConfig
+): Promise<ReduxStoreResult> {
     const { sessionUid, userId, masterKey, projectName } = config;
 
     logger.info({ userId: userId.slice(0, 8) + '...' }, 'Initializing upstream storage');
@@ -164,7 +283,6 @@ async function waitForReduxLoaded(
  * The initAppSaga triggers pullSpacesRequest after loading from IDB.
  * We need to wait for that to complete before checking if our space exists,
  * otherwise we might create a local space that conflicts with a remote one.
- * TODO: this looks like a good thing to have on a generic level
  */
 async function waitForRemoteSpaces(
     store: LumoStore,
@@ -278,7 +396,7 @@ function findOrCreateSpace(
  * Note: This does NOT handle messages that fail to decrypt (e.g., key mismatch
  * or corruption). Those remain in IDB but are skipped during loadReduxFromIdb.
  */
-export async function pullIncompleteConversations(
+async function pullIncompleteConversations(
     store: LumoStore,
     spaceId: SpaceId
 ): Promise<void> {
